Move Menus dropdown path
